# specs/002-pdp-product/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PDP P1 平台 API
  version: 1.1.0
  description: |
    PDP 企业级项目交付管理平台 P1 的稳定同步接口边界。
    本契约与 coverage.md 共同覆盖全部 P1 稳定操作；领域包生成页面使用
    domain-package.schema.json 定义的元数据，不为客户复制私有接口。
...
servers:
  - url: /api/v1
security:
  - oidc: []
...
paths:
...
  /domain-packages:
    get:
      tags: [领域包]
      summary: 查询当前工作空间可用领域包
      operationId: listDomainPackages
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - name: layer
          in: query
          schema:
            type: string
            enum: [PLATFORM_STANDARD, INDUSTRY, WORKSPACE_CUSTOMER]
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 领域包分页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainPackagePage'
    post:
      tags: [领域包]
      summary: 创建领域包草稿
      operationId: createDomainPackage
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDomainPackageCommand'
      responses:
        '201':
          description: 已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainPackage'
  /domain-packages/{packageId}/versions:
    post:
      tags: [领域包]
      summary: 创建领域包版本草稿
      operationId: createDomainPackageVersion
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/PackageIdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDomainPackageVersionCommand'
      responses:
        '201':
          description: 版本草稿已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainPackageVersion'
        '409':
          $ref: '#/components/responses/Problem'
  /domain-packages/{packageId}/versions/{versionId}/validate:
    post:
      tags: [领域包]
      summary: 校验领域包版本
      operationId: validateDomainPackageVersion
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/PackageIdPath'
        - $ref: '#/components/parameters/VersionIdPath'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: 校验完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainPackageValidationReport'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /domain-packages/{packageId}/versions/{versionId}/review:
    post:
      tags: [领域包]
      summary: 审核领域包版本
      operationId: reviewDomainPackageVersion
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/PackageIdPath'
        - $ref: '#/components/parameters/VersionIdPath'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewDomainPackageCommand'
      responses:
        '200':
          description: 审核完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainPackageVersion'
        '409':
          $ref: '#/components/responses/Problem'
  /domain-packages/{packageId}/versions/{versionId}/publish:
    post:
      tags: [领域包]
      summary: 审核并发布领域包版本
      operationId: publishDomainPackageVersion
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/PackageIdPath'
        - $ref: '#/components/parameters/VersionIdPath'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishPackageCommand'
      responses:
        '200':
          description: 已发布
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainPackageVersion'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /domain-packages/{packageId}/versions/{versionId}/retire:
    post:
      tags: [领域包]
      summary: 退役领域包版本
      operationId: retireDomainPackageVersion
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/PackageIdPath'
        - $ref: '#/components/parameters/VersionIdPath'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReasonCommand'
      responses:
        '200':
          description: 已退役
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainPackageVersion'
  /domain-packages/{packageId}/versions/{versionId}/rollback:
    post:
      tags: [领域包]
      summary: 回滚领域包版本
      operationId: rollbackDomainPackageVersion
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/PackageIdPath'
        - $ref: '#/components/parameters/VersionIdPath'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RollbackDomainPackageCommand'
      responses:
        '200':
          description: 已回滚
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainPackageVersion'
        '409':
          $ref: '#/components/responses/Problem'
  /domain-packages/{packageId}/versions/{versionId}/migration-preview:
    post:
      tags: [领域包]
      summary: 预览领域包升级影响
      operationId: previewDomainPackageMigration
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/PackageIdPath'
        - $ref: '#/components/parameters/VersionIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreviewDomainPackageMigrationCommand'
      responses:
        '200':
          description: 影响预览
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainPackageMigrationPreview'
  /domain-packages/{packageId}/versions/{versionId}/migrations:
    post:
      tags: [领域包]
      summary: 启动领域包实例迁移
      operationId: startDomainPackageMigration
      parameters:
        - $ref: '#/components/parameters/WorkspaceHeader'
        - $ref: '#/components/parameters/PackageIdPath'
        - $ref: '#/components/parameters/VersionIdPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartDomainPackageMigrationCommand'
      responses:
        '202':
          description: 迁移作业已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainPackageMigrationJob'
        '409':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://idp.example.com/.well-known/openid-configuration
  parameters:
    WorkspaceHeader:
      name: X-Workspace-Id
      in: header
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
    PackageIdPath:
      name: packageId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
    VersionIdPath:
      name: versionId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: 当前资源 revision 对应的强校验值；不匹配返回 409
      schema:
        type: string
    Cursor:
      name: cursor
      in: query
      description: 平台签发的不透明 keyset 游标，绑定排序、过滤和权限范围；禁止客户端解析或修改
      schema:
        type: string
        minLength: 16
        maxLength: 2048
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    Problem:
      description: 标准问题响应
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Uuid:
      type: string
      format: uuid
...
    Problem:
      type: object
      required: [type, title, status, traceId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
        traceId:
          type: string
        violations:
          type: array
          items:
            type: object
            required: [field, code]
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
...
    ReasonCommand:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 2
          maxLength: 1000
    CreateDomainPackageCommand:
      type: object
      required: [stableKey, name, layer]
      properties:
        stableKey:
          type: string
          pattern: '^[a-z][a-z0-9.-]{2,99}$'
        name:
          type: string
          maxLength: 100
        layer:
          type: string
          enum: [PLATFORM_STANDARD, INDUSTRY, WORKSPACE_CUSTOMER]
        parentPackageId:
          oneOf:
            - $ref: '#/components/schemas/Uuid'
            - type: 'null'
    DomainPackage:
      allOf:
        - $ref: '#/components/schemas/CreateDomainPackageCommand'
        - type: object
          required: [id, status]
          properties:
            id:
              $ref: '#/components/schemas/Uuid'
            status:
              type: string
              enum: [DRAFT, ACTIVE, DEPRECATED, RETIRED]
    CreateDomainPackageVersionCommand:
      type: object
      required: [semanticVersion, manifest]
      properties:
        semanticVersion:
          type: string
          pattern: '^[0-9]+\.[0-9]+\.[0-9]+$'
        manifest:
          type: object
          description: 必须符合 domain-package.schema.json
          additionalProperties: true
        changeSummary:
          type: string
          maxLength: 2000
    DomainPackageVersion:
      type: object
      required: [id, packageId, semanticVersion, contentHash, status, frozen, revision]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        packageId:
          $ref: '#/components/schemas/Uuid'
        semanticVersion:
          type: string
          pattern: '^[0-9]+\.[0-9]+\.[0-9]+$'
        contentHash:
          type: string
        status:
          type: string
          enum: [DRAFT, VALIDATED, REVIEW_PENDING, APPROVED, PUBLISHED, RETIRED, ROLLED_BACK]
        frozen:
          type: boolean
        revision:
          type: integer
          minimum: 0
    DomainPackagePage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DomainPackage'
        nextCursor:
          type: [string, 'null']
          description: 下一页不透明 keyset 游标；为 null 表示没有下一页
    PublishPackageCommand:
      type: object
      required: [reviewComment]
      properties:
        reviewComment:
          type: string
          minLength: 2
          maxLength: 1000
    ReviewDomainPackageCommand:
      type: object
      required: [decision, comment]
      properties:
        decision:
          type: string
          enum: [APPROVE, REJECT]
        comment:
          type: string
          minLength: 2
          maxLength: 1000
    RollbackDomainPackageCommand:
      type: object
      required: [targetVersionId, reason]
      properties:
        targetVersionId:
          $ref: '#/components/schemas/Uuid'
        reason:
          type: string
          minLength: 2
          maxLength: 1000
    DomainPackageValidationReport:
      type: object
      required: [valid, errors, warnings, compatibility]
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        compatibility:
          type: string
          enum: [COMPATIBLE, MIGRATION_REQUIRED, BREAKING]
    PreviewDomainPackageMigrationCommand:
      type: object
      required: [sourceVersionId, scope]
      properties:
        sourceVersionId:
          $ref: '#/components/schemas/Uuid'
        scope:
          type: object
          additionalProperties: true
        batchSize:
          type: integer
          minimum: 1
          maximum: 1000
          default: 100
    DomainPackageMigrationPreview:
      type: object
      required: [previewId, affectedInstances, conflicts, batches, rollbackAvailable, confirmationToken]
      properties:
        previewId:
          $ref: '#/components/schemas/Uuid'
        affectedInstances:
          type: integer
          minimum: 0
        conflicts:
          type: array
          items:
            type: string
        batches:
          type: integer
          minimum: 0
        rollbackAvailable:
          type: boolean
        confirmationToken:
          type: string
    StartDomainPackageMigrationCommand:
      type: object
      required: [previewId, confirmationToken, batchSize, reason]
      properties:
        previewId:
          $ref: '#/components/schemas/Uuid'
        confirmationToken:
          type: string
        batchSize:
          type: integer
          minimum: 1
          maximum: 1000
        reason:
          type: string
          minLength: 2
          maxLength: 1000
    DomainPackageMigrationJob:
      type: object
      required: [id, sourceVersionId, targetVersionId, status, migrated, failed, revision]
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        sourceVersionId:
          $ref: '#/components/schemas/Uuid'
        targetVersionId:
          $ref: '#/components/schemas/Uuid'
        status:
          type: string
          enum: [PLANNED, RUNNING, PARTIALLY_FAILED, COMPLETED, ROLLING_BACK, ROLLED_BACK, FAILED]
        migrated:
          type: integer
          minimum: 0
        failed:
          type: integer
          minimum: 0
        revision:
          type: integer
          minimum: 0
...